Resolve error on session in questions controller
There seems to have been an application level error report over
the last few days when users login or return to question pages
after time away. I suspected this was because the session had been
cleared.

I have therefore added a checkout on the show, answer and results action
in the questions controller to redirect to dashboard if session does
not contain any question info